+++ requirements.txt
@@ -7,3 +7,4 @@ pydot
 pyformlang
 pytest==7.1.3
 scipy
+numpy
