+++ tests/test_matrix_cfpq.py
@@ -15,14 +15,14 @@ def generate_graph(nodes: set[int], edges: set[tuple[int, str, int]]) -> MultiDi
 path = "resources/cfg_files"
 
 
-class TestsForHellingsCfpq:
+class TestsForMatrixCfpq:
     def test_works_as_expected_1(self):
         graph = generate_graph(
             {0, 1, 2, 3, 4, 5},
             {(0, "m", 1), (1, "i", 2), (2, "s", 3), (3, "h", 4), (4, "a", 5)},
         )
         cfg = cfg_from_file(f"{path}/6.txt")
-        res = cfpq(cfg, graph, {0}, {5}, algorithm=CfpqAlgorithms.HELLINGS)
+        res = cfpq(cfg, graph, algorithm=CfpqAlgorithms.MATRIX)
 
         assert res == {(0, 5)}
 
@@ -33,8 +33,8 @@ class TestsForHellingsCfpq:
         )
         cfg = cfg_from_file(f"{path}/7.txt")
 
-        res = cfpq(cfg, graph, start_nodes={3}, algorithm=CfpqAlgorithms.HELLINGS)
+        res = cfpq(cfg, graph, start_nodes={3}, algorithm=CfpqAlgorithms.MATRIX)
         assert res == set()
 
-        res = cfpq(cfg, graph, start_nodes={0}, final_nodes={2, 3}, algorithm=CfpqAlgorithms.HELLINGS)
+        res = cfpq(cfg, graph, start_nodes={0}, final_nodes={2, 3}, algorithm=CfpqAlgorithms.MATRIX)
         assert res == {(0, 2), (0, 3)}
